python/requirements.txt
streamlit
pymongo
python-dotenv
langchain-openai
pandas
plotly
fastapi
uvicorn
python-dotenv
pymongo
requests
